refactor: DNSConfig等の各種configを使用する側の.goで保持するよう調整
